requirements.txt
click==8.1.1
Flask==2.1.1
Flask-JWT-Extended==4.3.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dotenv==0.20.0
SQLAlchemy==1.4.34
Werkzeug==2.1.0
zipp==3.7.0
